% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chooser.R
\name{mod_chooser_ui}
\alias{mod_chooser_ui}
\alias{mod_chooser_server}
\title{mod_chooser_ui and mod_chooser_server}
\usage{
mod_chooser_ui(id)

mod_chooser_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
