% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datamaps.R
\name{mod_datamaps_ui}
\alias{mod_datamaps_ui}
\alias{mod_datamaps_server}
\title{mod_datamaps_ui and mod_datamaps_server}
\usage{
mod_datamaps_ui(id)

mod_datamaps_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
