% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap.R
\name{mod_heatmap_ui}
\alias{mod_heatmap_ui}
\alias{mod_heatmap_server}
\title{mod_heatmap_ui and mod_heatmap_server}
\usage{
mod_heatmap_ui(id)

mod_heatmap_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
