% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_leaftime.R
\name{mod_leaftime_ui}
\alias{mod_leaftime_ui}
\alias{mod_leaftime_server}
\title{mod_leaftime_ui and mod_leaftime_server}
\usage{
mod_leaftime_ui(id)

mod_leaftime_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
