% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapview.R
\name{mod_mapview_ui}
\alias{mod_mapview_ui}
\alias{mod_mapview_server}
\title{mod_mapview_ui and mod_mapview_server}
\usage{
mod_mapview_ui(id)

mod_mapview_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
