% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phylo.R
\name{mod_phylo_ui}
\alias{mod_phylo_ui}
\alias{mod_phylo_server}
\title{mod_phylo_ui and mod_phylo_server}
\usage{
mod_phylo_ui(id)

mod_phylo_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
