% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scroll.R
\name{mod_scroll_ui}
\alias{mod_scroll_ui}
\alias{mod_scroll_server}
\title{mod_scroll_ui and mod_scroll_server}
\usage{
mod_scroll_ui(id)

mod_scroll_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
