% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_webVr.R
\name{mod_webVr_ui}
\alias{mod_webVr_ui}
\alias{mod_webVr_server}
\title{mod_webVr_ui and mod_webVr_server}
\usage{
mod_webVr_ui(id)

mod_webVr_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
