% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yearly_chart.R
\name{mod_yearly_chart_ui}
\alias{mod_yearly_chart_ui}
\alias{mod_yearly_chart_server}
\title{mod_yearly_chart_ui and mod_yearly_chart_server}
\usage{
mod_yearly_chart_ui(id)

mod_yearly_chart_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
